tests/wpt/web-platform-tests/shadow-dom/html-elements-and-their-shadow-trees/test-002.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_09_00_02</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#html-elements-and-their-shadow-trees">
<meta name="assert" content="HTML Elements and Their Shadow Trees: Elements that have no fallback content should allow the shadow tree to contain no insertion points or an insertion point that matches nothing">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
//test img
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('img');
    d.body.appendChild(el);

    var s = el.createShadowRoot();

    s.innerHTML = '<content id="cont" select="#shadow"></content>';

    assert_true(s.querySelector('#cont') != null, 'img should allow one insertion point ' +
    		'that matches nothing');

}), 'A_09_00_02_T01');


//test embed
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('embed');
    d.body.appendChild(el);

    var s = el.createShadowRoot();

    s.innerHTML = '<content id="cont" select="#shadow"></content>';

    assert_true(s.querySelector('#cont') != null, 'embed should allow one insertion point ' +
    		'that matches nothing');

}), 'A_09_00_02_T02');


//test embed
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	// create element
    var el = d.createElement('input');
    d.body.appendChild(el);

    var s = el.createShadowRoot();

    s.innerHTML = '<content id="cont" select="#shadow"></content>';

    assert_true(s.querySelector('#cont') != null, 'input should allow one insertion point ' +
    		'that matches nothing');

}), 'A_09_00_02_T03');
</script>
</body>
</html>
